Flatten: move join-split up
