+++ barretenberg/cpp/src/barretenberg/join_split_example/CMakeLists.txt
@@ -1,2 +1,12 @@
-add_subdirectory(join_split)
-add_subdirectory(notes)
+barretenberg_module(
+  join_split_example_proofs_join_split
+  join_split_example_proofs_notes
+  crypto_schnorr
+  stdlib_blake2s
+  stdlib_sha256
+  stdlib_pedersen_commitment
+  stdlib_schnorr
+  stdlib_primitives
+  crypto_merkle_tree)
+  
+  add_subdirectory(notes)
