+++ barretenberg/cpp/src/barretenberg/join_split_example/compute_signing_data.hpp
