+++ barretenberg/cpp/src/barretenberg/join_split_example/inner_proof_data.hpp
