+++ barretenberg/cpp/src/barretenberg/join_split_example/notes/circuit/claim/compute_nullifier.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../../constants.hpp"
+#include "barretenberg/join_split_example/constants.hpp"
 #include "barretenberg/join_split_example/types.hpp"
 #include "barretenberg/stdlib/hash/pedersen/pedersen.hpp"
