+++ barretenberg/cpp/src/barretenberg/join_split_example/notes/circuit/claim/create_partial_commitment.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "../../constants.hpp"
+#include "barretenberg/join_split_example/constants.hpp"
 #include "barretenberg/join_split_example/types.hpp"
 #include "barretenberg/stdlib/hash/pedersen/pedersen.hpp"
