+++ barretenberg/cpp/src/barretenberg/join_split_example/notes/circuit/value/compute_nullifier.cpp
@@ -1,5 +1,5 @@
 #include "compute_nullifier.hpp"
-#include "../../constants.hpp"
+#include "barretenberg/join_split_example/notes/constants.hpp"
 #include "barretenberg/join_split_example/types.hpp"
 
 namespace bb::join_split_example::proofs::notes::circuit {
