+++ barretenberg/cpp/src/barretenberg/join_split_example/notes/circuit/value/compute_nullifier.test.cpp
@@ -1,9 +1,9 @@
-#include "../../native/value/compute_nullifier.hpp"
-#include "../../../../fixtures/user_context.hpp"
-#include "../../native/value/value_note.hpp"
+#include "barretenberg/join_split_example/notes/native/value/compute_nullifier.hpp"
 #include "./compute_nullifier.hpp"
 #include "./value_note.hpp"
+#include "barretenberg/join_split_example/notes/native/value/value_note.hpp"
 #include "barretenberg/join_split_example/types.hpp"
+#include "barretenberg/join_split_example/user_context.hpp"
 #include <gtest/gtest.h>
 
 namespace bb::join_split_example {
