+++ barretenberg/cpp/src/barretenberg/join_split_example/notes/native/account/account_note.cpp
@@ -1,6 +1,6 @@
 #include "account_note.hpp"
-#include "../../constants.hpp"
 #include "barretenberg/crypto/pedersen_hash/pedersen.hpp"
+#include "barretenberg/join_split_example/notes/constants.hpp"
 
 namespace bb::join_split_example::proofs::notes::native::account {
