+++ barretenberg/cpp/src/barretenberg/join_split_example/notes/native/account/account_note.hpp
@@ -1,7 +1,7 @@
 #pragma once
-#include "../../constants.hpp"
 #include "barretenberg/crypto/pedersen_commitment/pedersen.hpp"
 #include "barretenberg/ecc/curves/grumpkin/grumpkin.hpp"
+#include "barretenberg/join_split_example/constants.hpp"
 
 namespace bb::join_split_example::proofs::notes::native::account {
