+++ barretenberg/cpp/src/barretenberg/join_split_example/notes/native/claim/claim_note_tx_data.hpp
@@ -1,8 +1,8 @@
 #pragma once
-#include "../bridge_call_data.hpp"
 #include "barretenberg/common/serialize.hpp"
 #include "barretenberg/crypto/pedersen_commitment/pedersen.hpp"
 #include "barretenberg/ecc/curves/grumpkin/grumpkin.hpp"
+#include "barretenberg/join_split_example/notes/native/bridge_call_data.hpp"
 
 namespace bb::join_split_example::proofs::notes::native::claim {
