+++ barretenberg/cpp/src/barretenberg/join_split_example/notes/native/value/complete_partial_commitment.hpp
@@ -1,6 +1,6 @@
 #pragma once
-#include "../../constants.hpp"
 #include "barretenberg/crypto/pedersen_hash/pedersen.hpp"
+#include "barretenberg/join_split_example/notes/constants.hpp"
 
 namespace bb::join_split_example::proofs::notes::native::value {
