+++ barretenberg/cpp/src/barretenberg/join_split_example/notes/native/value/compute_nullifier.cpp
@@ -1,8 +1,8 @@
 #include "compute_nullifier.hpp"
-#include "../../constants.hpp"
 #include "barretenberg/crypto/blake2s/blake2s.hpp"
 #include "barretenberg/crypto/pedersen_commitment/pedersen.hpp"
 #include "barretenberg/crypto/pedersen_hash/pedersen.hpp"
+#include "barretenberg/join_split_example/notes/constants.hpp"
 
 namespace bb::join_split_example::proofs::notes::native {
