+++ barretenberg/cpp/src/barretenberg/join_split_example/sign_join_split_tx.cpp
