+++ barretenberg/cpp/src/barretenberg/join_split_example/verify_signature.hpp
